+++ app/assets/javascripts/modules/comment.js
