+++ app/assets/stylesheets/application.scss
@@ -3,6 +3,7 @@
 @import "font-awesome-sprockets";
 @import "font-awesome";
 @import "modules/variables";
+@import "fullcalendar";
 // @import "simple_calendar";
 @import "modules/calendar";
 @import "modules/editCalendar";
