+++ app/javascript/packs/application.js
@@ -7,8 +7,10 @@ require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-require("calendar")
-require("comment")
+
+
+// require("calendar")
+// require("comment")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
