+++ app/views/events/create.json.jbuilder
@@ -1,9 +1,9 @@
-json.id           @event.id
-json.title        @event.title
-json.content      @event.content
-json.strt_hour    @event.strt_hour.strftime("%H:%M")
-json.end_hour     @event.end_hour.strftime("%H:%M")
-json.start_day    @event.start_day.strftime("%Y-%m-%d")
-json.end_day      @event.end_day.strftime("%Y-%m-%d")
-json.user_name    @event.user.name
-json.calendar_id  @event.calendar.id
+# json.id           @event.id
+# json.title        @event.title
+# json.content      @event.content
+# json.strt_hour    @event.strt_hour.strftime("%H:%M")
+# json.end_hour     @event.end_hour.strftime("%H:%M")
+# json.start_day    @event.start_day.strftime("%Y-%m-%d")
+# json.end_day      @event.end_day.strftime("%Y-%m-%d")
+# json.user_name    @event.user.name
+# json.calendar_id  @event.calendar.id
