+++ app/views/layouts/application.html.haml
@@ -6,8 +6,8 @@
     = csrf_meta_tags
     = csp_meta_tag
     = stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload'
-    -# = javascript_include_tag 'application'
-    = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
-  %body
+    = javascript_include_tag 'application'
+    -# = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
+  %body{"data-turbolinks":"false"}
     = render 'layouts/notifications'
     = yield
